Bug 536324 part 3 - Change consumers of nsIChannel. r=sworkman
UltraBlame original commit: af7e6c8e6d768e7dcf2e9f7f30fc7f4c8b37d4f5

+++ dom/plugins/base/nsPluginStreamListenerPeer.cpp
@@ -556,7 +556,7 @@ nsPluginStreamListenerPeer::OnStartRequest(nsIRequest *request,
   if (loadGroup)
     mWeakPtrChannelLoadGroup = do_GetWeakReference(loadGroup);
   
-  int32_t length;
+  int64_t length;
   rv = channel->GetContentLength(&length);
